+++ src/test/java/org/folio/entlinks/controller/ConsortiumLinksSuggestionsIT.java
@@ -97,13 +97,14 @@ class ConsortiumLinksSuggestionsIT extends IntegrationTestBase {
 
   @Test
   @SneakyThrows
-  void getAuthDataStat_shouldActualizeLinkAndSubfieldsForMemberTenant() {
+  void suggestLinksForMarcRecord_shouldActualizeLinkAndSubfieldsForMemberTenant() {
     var givenSubfields = Map.of("a", "old $a value", "0", FULL_BASE_URL + NATURAL_ID);
     var givenLinkDetails = getLinkDetails(ACTUAL);
     var givenRecord = getRecord("100", givenLinkDetails, givenSubfields);
 
     var expectedLinkDetails = getLinkDetails(ACTUAL);
-    var expectedSubfields = Map.of("a", "new $a value", "0", FULL_BASE_URL + NATURAL_ID, "9", LINKABLE_AUTHORITY_ID);
+    var expectedSubfields = Map.of("a", "new $a value", "q", "new $q value", "b", "new $b value",
+      "0", FULL_BASE_URL + NATURAL_ID, "9", LINKABLE_AUTHORITY_ID);
     var expectedRecord = getRecord("100", expectedLinkDetails, expectedSubfields);
 
     var requestBody = new ParsedRecordContentCollection().records(List.of(givenRecord));
@@ -115,12 +116,13 @@ class ConsortiumLinksSuggestionsIT extends IntegrationTestBase {
 
   @Test
   @SneakyThrows
-  void getAuthDataStat_shouldSuggestNewLinkForMemberTenant() {
+  void suggestLinksForMarcRecord_shouldSuggestNewLinkForMemberTenant() {
     var givenSubfields = Map.of("0", NATURAL_ID);
     var givenRecord = getRecord("100", null, givenSubfields);
 
     var expectedLinkDetails = getLinkDetails(NEW);
-    var expectedSubfields = Map.of("a", "new $a value", "0", FULL_BASE_URL + NATURAL_ID, "9", LINKABLE_AUTHORITY_ID);
+    var expectedSubfields = Map.of("a", "new $a value", "q", "new $q value", "b", "new $b value",
+      "0", FULL_BASE_URL + NATURAL_ID, "9", LINKABLE_AUTHORITY_ID);
     var expectedRecord = getRecord("100", expectedLinkDetails, expectedSubfields);
 
     var requestBody = new ParsedRecordContentCollection().records(List.of(givenRecord));
@@ -132,12 +134,13 @@ class ConsortiumLinksSuggestionsIT extends IntegrationTestBase {
 
   @Test
   @SneakyThrows
-  void getAuthDataStat_shouldSuggestNewLinkByAuthorityIdForMemberTenant() {
+  void suggestLinksForMarcRecord_shouldSuggestNewLinkByAuthorityIdForMemberTenant() {
     var givenSubfields = Map.of("9", LINKABLE_AUTHORITY_ID);
     var givenRecord = getRecord("100", null, givenSubfields);
 
     var expectedLinkDetails = getLinkDetails(NEW);
-    var expectedSubfields = Map.of("a", "new $a value", "0", FULL_BASE_URL + NATURAL_ID, "9", LINKABLE_AUTHORITY_ID);
+    var expectedSubfields = Map.of("a", "new $a value", "q", "new $q value", "b", "new $b value",
+      "0", FULL_BASE_URL + NATURAL_ID, "9", LINKABLE_AUTHORITY_ID);
     var expectedRecord = getRecord("100", expectedLinkDetails, expectedSubfields);
 
     var requestBody = new ParsedRecordContentCollection().records(List.of(givenRecord));
@@ -149,7 +152,7 @@ class ConsortiumLinksSuggestionsIT extends IntegrationTestBase {
 
   @Test
   @SneakyThrows
-  void getAuthDataStat_shouldFillErrorDetails_whenNoSuggestionsFound() {
+  void suggestLinksForMarcRecord_shouldFillErrorDetails_whenNoSuggestionsFound() {
     var givenSubfields = Map.of("0", NATURAL_ID);
     var givenRecord = getRecord("110", null, givenSubfields);